+++ clients/cpp/src/extreme_carpaccio_client/ExtremeCarpaccioClient.cpp
@@ -249,19 +249,5 @@ void http_worker::check_deadline()
       http::write(m_stream, request);
    }
    
-   EXTREME_CARPACCIO_CLIENT_API Order parseOrder(const std::string & jsonOrder)
-   {
-      Order order;
-
-      auto parsedOrder = nlohmann::json::parse(jsonOrder);
-
-      order.country = parsedOrder["country"].get<std::string>();
-      order.reduction = parsedOrder["reduction"].get<std::string>();
-      order.prices = parsedOrder["prices"].get<std::vector<double>>();
-      order.quantities = parsedOrder["quantities"].get<std::vector<unsigned int>>();
-
-      return order;
-   }
-
 } // namespace extreme_carpaccio_client
